* Cleanup of MemberWorker and some related classes. * Pulled TestContextImpl, WorkerCommandRequestProcessor and WorkerSocketProcessor out of MemberWorker. * Hid thread of PerformanceMonitor within class. Renamed PerformanceMonitor to WorkerPerformanceMonitor. WorkerPerformanceMonitor is now lazily started at first run command and stopped after run phase. So performance.txt files don't contain empty data from warmup and verify phases anymore. Prevented creation of files for tests without performance monitoring.
