+++ stabilizer/src/main/java/com/hazelcast/stabilizer/worker/ClientWorker.java
@@ -1,10 +1,8 @@
 package com.hazelcast.stabilizer.worker;
 
-
 /**
- * This class serves no other purpose that to provide a name in the jps listing that reflects that the jvm is a client
- * jvm. It has no other purpose and it will delegate all its work to the {@link MemberWorker}
- * class.
+ * This class serves no other purpose than to provide a name in the jps listing that reflects that the JVM is a client JVM.
+ * It has no other purpose and it will delegate all its work to the {@link MemberWorker} class.
  */
 public class ClientWorker {
