+++ stabilizer/src/main/java/com/hazelcast/stabilizer/worker/IllegalTestException.java
@@ -1,7 +1,8 @@
 package com.hazelcast.stabilizer.worker;
 
 /**
- * Exception thrown when a Test is not valid, e.g. it has no method with a @Run annotation.
+ * Exception thrown when a test is not valid, e.g. it has no method with a {@link com.hazelcast.stabilizer.test.annotations.Run}
+ * or {@link com.hazelcast.stabilizer.test.annotations.RunWithWorker} annotation.
  */
 public class IllegalTestException extends RuntimeException {
