+++ stabilizer/src/main/java/com/hazelcast/stabilizer/worker/MemberWorker.java
@@ -22,72 +22,34 @@ import com.hazelcast.config.Config;
 import com.hazelcast.config.XmlConfigBuilder;
 import com.hazelcast.core.Hazelcast;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.stabilizer.agent.workerjvm.WorkerJvmManager;
-import com.hazelcast.stabilizer.common.messaging.Message;
-import com.hazelcast.stabilizer.probes.probes.ProbesConfiguration;
-import com.hazelcast.stabilizer.probes.probes.Result;
-import com.hazelcast.stabilizer.test.TestCase;
-import com.hazelcast.stabilizer.test.TestContext;
 import com.hazelcast.stabilizer.test.utils.TestUtils;
 import com.hazelcast.stabilizer.utils.ExceptionReporter;
-import com.hazelcast.stabilizer.worker.commands.Command;
 import com.hazelcast.stabilizer.worker.commands.CommandRequest;
 import com.hazelcast.stabilizer.worker.commands.CommandResponse;
-import com.hazelcast.stabilizer.worker.commands.GenericCommand;
-import com.hazelcast.stabilizer.worker.commands.GetBenchmarkResultsCommand;
-import com.hazelcast.stabilizer.worker.commands.GetOperationCountCommand;
-import com.hazelcast.stabilizer.worker.commands.InitCommand;
-import com.hazelcast.stabilizer.worker.commands.IsPhaseCompletedCommand;
-import com.hazelcast.stabilizer.worker.commands.MessageCommand;
-import com.hazelcast.stabilizer.worker.commands.RunCommand;
-import com.hazelcast.stabilizer.worker.commands.StopCommand;
 import org.apache.log4j.LogManager;
 import org.apache.log4j.Logger;
 
 import java.io.File;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.lang.management.ManagementFactory;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
-import java.net.Socket;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.TimeUnit;
 
-import static com.hazelcast.stabilizer.utils.CommonUtils.closeQuietly;
-import static com.hazelcast.stabilizer.utils.CommonUtils.fixRemoteStackTrace;
 import static com.hazelcast.stabilizer.utils.CommonUtils.getHostAddress;
 import static com.hazelcast.stabilizer.utils.FileUtils.fileAsText;
-import static com.hazelcast.stabilizer.utils.FileUtils.isValidFileName;
 import static com.hazelcast.stabilizer.utils.FileUtils.writeObject;
-import static com.hazelcast.stabilizer.utils.PropertyBindingSupport.bindProperties;
-import static com.hazelcast.stabilizer.utils.PropertyBindingSupport.parseProbeConfiguration;
 import static java.lang.String.format;
-import static java.util.Arrays.asList;
 
 public class MemberWorker {
 
-    private static final String DASHES = "---------------------------";
     private static final Logger log = Logger.getLogger(MemberWorker.class);
-
-    private final ConcurrentMap<String, Command> commands = new ConcurrentHashMap<String, Command>();
-    private final ConcurrentMap<String, TestContainer<TestContext>> tests = new ConcurrentHashMap<String, TestContainer<TestContext>>();
-
-    private final WorkerMessageProcessor workerMessageProcessor = new WorkerMessageProcessor(tests);
 
     private final BlockingQueue<CommandRequest> requestQueue = new LinkedBlockingQueue<CommandRequest>();
     private final BlockingQueue<CommandResponse> responseQueue = new LinkedBlockingQueue<CommandResponse>();
 
-    private HazelcastInstance serverInstance;
-    private HazelcastInstance clientInstance;
+    private WorkerSocketProcessor workerSocketProcessor;
+    private WorkerCommandRequestProcessor workerCommandRequestProcessor;
 
     private String hzFile;
     private String clientHzFile;
@@ -95,70 +57,49 @@ public class MemberWorker {
     private String workerMode;
     private String workerId;
 
-    public void start() throws Exception {
-        if ("server".equals(workerMode)) {
-            log.info("------------------------------------------------------------------------");
-            log.info("             member mode");
-            log.info("------------------------------------------------------------------------");
-            this.serverInstance = createServerHazelcastInstance();
-            TestUtils.warmupPartitions(log, serverInstance);
-        } else if ("client".equals(workerMode)) {
-            log.info("------------------------------------------------------------------------");
-            log.info("             client mode");
-            log.info("------------------------------------------------------------------------");
-            this.clientInstance = createClientHazelcastInstance();
-            TestUtils.warmupPartitions(log, clientInstance);
-        } else {
-            throw new IllegalStateException("Unknown worker mode:" + workerMode);
+    public static void main(String[] args) {
+        log.info("Starting Stabilizer Worker");
+
+        try {
+            logInputArguments();
+            logInterestingSystemProperties();
+
+            String workerId = System.getProperty("workerId");
+            log.info("Worker id: " + workerId);
+
+            String workerHzFile = args[0];
+            log.info("Worker hz config file: " + workerHzFile);
+            log.info(fileAsText(new File(workerHzFile)));
+
+            String clientHzFile = args[1];
+            log.info("Client hz config file: " + clientHzFile);
+            log.info(fileAsText(new File(clientHzFile)));
+
+            String workerMode = System.getProperty("workerMode");
+            log.info("Worker mode: " + workerMode);
+
+            MemberWorker worker = new MemberWorker();
+            worker.workerId = workerId;
+            worker.hzFile = workerHzFile;
+            worker.clientHzFile = clientHzFile;
+            worker.workerMode = workerMode;
+            worker.start();
+
+            registerLog4jShutdownHandler(worker);
+
+            log.info("Successfully started Hazelcast Stabilizer Worker: " + workerId);
+        } catch (Throwable e) {
+            ExceptionReporter.report(null, e);
+            System.exit(1);
         }
-        log.info("------------------------------------------------------------------------");
-
-        workerMessageProcessor.setHazelcastServerInstance(serverInstance);
-        workerMessageProcessor.setHazelcastClientInstance(clientInstance);
-
-        new CommandRequestProcessingThread().start();
-        new SocketThread().start();
-        new PerformanceMonitor(tests.values()).start();
-
-        // the last thing we do is to signal to the agent we have started.
-        signalStartToAgent();
     }
 
-    private void signalStartToAgent() {
-        String address;
-        if (serverInstance == null) {
-            address = "client:" + getHostAddress();
-        } else {
-            InetSocketAddress socketAddress = serverInstance.getCluster().getLocalMember().getInetSocketAddress();
-            address = socketAddress.getAddress().getHostAddress() + ":" + socketAddress.getPort();
-        }
-        File file = new File("worker.address");
-        writeObject(address, file);
+    private static void logInputArguments() {
+        List<String> inputArguments = ManagementFactory.getRuntimeMXBean().getInputArguments();
+        log.info("JVM input arguments: " + inputArguments);
     }
 
-    private HazelcastInstance createClientHazelcastInstance() throws Exception {
-        log.info("Creating Client HazelcastInstance");
-
-        XmlClientConfigBuilder configBuilder = new XmlClientConfigBuilder(clientHzFile);
-        ClientConfig clientConfig = configBuilder.build();
-
-        HazelcastInstance client = HazelcastClient.newHazelcastClient(clientConfig);
-        log.info("Successfully created Client HazelcastInstance");
-        return client;
-    }
-
-    private HazelcastInstance createServerHazelcastInstance() throws Exception {
-        log.info("Creating Server HazelcastInstance");
-
-        XmlConfigBuilder configBuilder = new XmlConfigBuilder(hzFile);
-        Config config = configBuilder.build();
-
-        HazelcastInstance server = Hazelcast.newHazelcastInstance(config);
-        log.info("Successfully created Server HazelcastInstance");
-        return server;
-    }
-
-    protected static void logInterestingSystemProperties() {
+    private static void logInterestingSystemProperties() {
         logSystemProperty("java.class.path");
         logSystemProperty("java.home");
         logSystemProperty("java.vendor");
@@ -180,386 +121,85 @@ public class MemberWorker {
         log.info(format("%s=%s", name, System.getProperty(name)));
     }
 
-    public static void main(String[] args) {
-        registerLog4jShutdownHandler();
-
-        log.info("Starting Stabilizer Worker");
-
-        try {
-            logInputArguments();
-            logInterestingSystemProperties();
-
-            String workerId = System.getProperty("workerId");
-            log.info("Worker id:" + workerId);
-
-            String workerHzFile = args[0];
-            log.info("Worker hz config file:" + workerHzFile);
-            log.info(fileAsText(new File(workerHzFile)));
-
-            String clientHzFile = args[1];
-            log.info("Client hz config file:" + clientHzFile);
-            log.info(fileAsText(new File(clientHzFile)));
-
-            String workerMode = System.getProperty("workerMode");
-            log.info("Worker mode:" + workerMode);
-
-            MemberWorker worker = new MemberWorker();
-            worker.workerId = workerId;
-            worker.hzFile = workerHzFile;
-            worker.clientHzFile = clientHzFile;
-            worker.workerMode = workerMode;
-            worker.start();
-
-            log.info("Successfully started Hazelcast Stabilizer Worker:" + workerId);
-        } catch (Throwable e) {
-            ExceptionReporter.report(null, e);
-            System.exit(1);
-        }
-    }
-
-    private static void registerLog4jShutdownHandler() {
+    private static void registerLog4jShutdownHandler(final MemberWorker worker) {
         // makes sure that log4j will always flush log-buffers
         Runtime.getRuntime().addShutdownHook(new Thread() {
             @Override
             public void run() {
+                worker.stop();
                 LogManager.shutdown();
             }
         });
     }
 
-    protected static void logInputArguments() {
-        List<String> inputArguments = ManagementFactory.getRuntimeMXBean().getInputArguments();
-        log.info("jvm input arguments = " + inputArguments);
+    private MemberWorker() {
     }
 
-    private class SocketThread extends Thread {
+    private void start() throws Exception {
+        HazelcastInstance serverInstance = null;
+        HazelcastInstance clientInstance = null;
 
-        @Override
-        public void run() {
-            for (; ; ) {
-                try {
-                    List<CommandRequest> requests = execute(WorkerJvmManager.SERVICE_POLL_WORK, workerId);
-                    for (CommandRequest request : requests) {
-                        requestQueue.add(request);
-                    }
-
-                    CommandResponse response = responseQueue.poll(1, TimeUnit.SECONDS);
-                    if (response == null) {
-                        continue;
-                    }
-
-                    sendResponse(asList(response));
-
-                    List<CommandResponse> responses = new LinkedList<CommandResponse>();
-                    responseQueue.drainTo(responses);
-
-                    sendResponse(responses);
-                } catch (Throwable e) {
-                    ExceptionReporter.report(null, e);
-                }
-            }
+        if ("server".equals(workerMode)) {
+            log.info("------------------------------------------------------------------------");
+            log.info("             member mode");
+            log.info("------------------------------------------------------------------------");
+            serverInstance = createServerHazelcastInstance();
+            TestUtils.warmupPartitions(log, serverInstance);
+        } else if ("client".equals(workerMode)) {
+            log.info("------------------------------------------------------------------------");
+            log.info("             client mode");
+            log.info("------------------------------------------------------------------------");
+            clientInstance = createClientHazelcastInstance();
+            TestUtils.warmupPartitions(log, clientInstance);
+        } else {
+            throw new IllegalStateException("Unknown worker mode: " + workerMode);
         }
 
-        private void sendResponse(List<CommandResponse> responses) throws Exception {
-            for (CommandResponse response : responses) {
-                execute(WorkerJvmManager.COMMAND_PUSH_RESPONSE, workerId, response);
-            }
-        }
+        workerSocketProcessor = new WorkerSocketProcessor(requestQueue, responseQueue, workerId);
+        workerCommandRequestProcessor = new WorkerCommandRequestProcessor(requestQueue, responseQueue,
+                serverInstance, clientInstance);
 
-        // we create a new socket for every request because don't want to depend on the state of a socket
-        // since we are going to do nasty stuff.
-        private <E> E execute(String service, Object... args) throws Exception {
-            Socket socket = new Socket(InetAddress.getByName(null), WorkerJvmManager.PORT);
-
-            try {
-                ObjectOutputStream oos = new ObjectOutputStream(socket.getOutputStream());
-                oos.writeObject(service);
-                for (Object arg : args) {
-                    oos.writeObject(arg);
-                }
-                oos.flush();
-
-                ObjectInputStream in = new ObjectInputStream(socket.getInputStream());
-                Object response = in.readObject();
-
-                if (response instanceof TerminateWorkerException) {
-                    System.exit(0);
-                }
-
-                if (response instanceof Exception) {
-                    Exception exception = (Exception) response;
-                    fixRemoteStackTrace(exception, Thread.currentThread().getStackTrace());
-                    throw exception;
-                }
-
-                return (E) response;
-            } finally {
-                closeQuietly(socket);
-            }
-        }
+        // the last thing we do is to signal to the agent we have started
+        signalStartToAgent(serverInstance);
     }
 
-    private class CommandRequestProcessingThread extends Thread {
-
-        @Override
-        public void run() {
-            for (; ; ) {
-                try {
-                    CommandRequest request = requestQueue.take();
-                    if (request == null) {
-                        throw new NullPointerException("request can't be null");
-                    }
-                    doProcess(request.id, request.task);
-                } catch (Throwable e) {
-                    ExceptionReporter.report(null, e);
-                }
-            }
-        }
-
-        private void doProcess(long id, Command command) throws Throwable {
-            Object result = null;
-            try {
-                if (command instanceof IsPhaseCompletedCommand) {
-                    result = process((IsPhaseCompletedCommand) command);
-                } else if (command instanceof InitCommand) {
-                    process((InitCommand) command);
-                } else if (command instanceof RunCommand) {
-                    process((RunCommand) command);
-                } else if (command instanceof StopCommand) {
-                    process((StopCommand) command);
-                } else if (command instanceof GenericCommand) {
-                    process((GenericCommand) command);
-                } else if (command instanceof GetOperationCountCommand) {
-                    result = process((GetOperationCountCommand) command);
-                } else if (command instanceof GetBenchmarkResultsCommand) {
-                    result = process((GetBenchmarkResultsCommand) command);
-                } else if (command instanceof MessageCommand) {
-                    process((MessageCommand) command);
-                } else {
-                    throw new RuntimeException("Unhandled task:" + command.getClass());
-                }
-            } finally {
-                if (command.awaitReply()) {
-                    CommandResponse response = new CommandResponse();
-                    response.commandId = id;
-                    response.result = result;
-                    responseQueue.add(response);
-                }
-            }
-        }
-
-        private Map<String, Result<?>> process(GetBenchmarkResultsCommand command) {
-            String testId = command.getTestId();
-            return tests.get(testId).getProbeResults();
-        }
-
-        private void process(MessageCommand command) {
-            Message message = command.getMessage();
-            workerMessageProcessor.submit(message);
-        }
-
-        @SuppressWarnings("unused")
-        private Long process(GetOperationCountCommand command) throws Throwable {
-            long result = 0;
-
-            for (TestContainer testContainer : tests.values()) {
-                long operationCount = testContainer.getOperationCount();
-                if (operationCount > 0) {
-                    result += operationCount;
-                }
-            }
-
-            return result;
-        }
-
-        private void process(final RunCommand command) throws Exception {
-            try {
-                final String testId = command.testId;
-                final String testName = "".equals(testId) ? "test" : testId;
-
-                final TestContainer<TestContext> test = tests.get(testId);
-                if (test == null) {
-                    log.warn("Failed to process command: " + command + " no test with testId" + testId + " is found");
-                    return;
-                }
-
-                CommandThread commandThread = new CommandThread(command, testId) {
-                    @Override
-                    public void doRun() throws Throwable {
-                        boolean passive = command.clientOnly && clientInstance == null;
-
-                        if (passive) {
-                            log.info(format("%s Skipping %s.run() (member is passive) %s", DASHES, testName, DASHES));
-                        } else {
-                            log.info(format("%s Starting %s.run() %s", DASHES, testId, DASHES));
-
-                            try {
-                                test.run();
-                                log.info(format("%s Completed %s.run() %s", DASHES, testName, DASHES));
-                            } catch (InvocationTargetException e) {
-                                log.fatal(format("%s Failed to execute %s.run() %s", DASHES, testName, DASHES), e.getCause());
-                                throw e.getCause();
-                            }
-                        }
-                    }
-                };
-                commandThread.start();
-            } catch (Exception e) {
-                log.fatal("Failed to start test", e);
-                throw e;
-            }
-        }
-
-        public void process(final GenericCommand command) throws Throwable {
-            final String methodName = command.methodName;
-            final String testId = command.testId;
-            final String testName = "".equals(testId) ? "test" : testId;
-
-            try {
-                final TestContainer<TestContext> test = tests.get(testId);
-                if (test == null) {
-                    // we log a warning: it could be that it is a newly created machine from mama-monkey.
-                    log.warn("Failed to process command: " + command + " no test with " +
-                            "testId " + testId + " is found");
-                    return;
-                }
-
-                final Method method = test.getClass().getMethod(methodName);
-                CommandThread commandThread = new CommandThread(command, command.testId) {
-                    @Override
-                    public void doRun() throws Throwable {
-                        log.info(format("%s Starting %s.%s() %s", DASHES, testName, methodName, DASHES));
-
-                        try {
-                            method.invoke(test);
-                            log.info(format("%s Finished %s.%s() %s", DASHES, testName, methodName, DASHES));
-                        } catch (InvocationTargetException e) {
-                            log.fatal(format("%s Failed %s.%s() %s", DASHES, testName, methodName, DASHES));
-                            throw e.getCause();
-                        } finally {
-                            if ("localTeardown".equals(methodName)) {
-                                tests.remove(testId);
-                            }
-                        }
-                    }
-                };
-                commandThread.start();
-            } catch (Exception e) {
-                log.fatal(format("Failed to execute test.%s()", methodName), e);
-                throw e;
-            }
-        }
-
-        private void process(InitCommand command) throws Throwable {
-            try {
-                TestCase testCase = command.testCase;
-                String testId = testCase.getId();
-                if (tests.containsKey(testId)) {
-                    throw new IllegalStateException(
-                            format("Can't init TestCase: %s, another test with testId [%s] already exists", command, testId));
-                }
-                if (!testId.isEmpty() && !isValidFileName(testId)) {
-                    throw new IllegalArgumentException(
-                            format("Can't init TestCase: %s, testId [%s] is an invalid filename", command, testId));
-                }
-
-                log.info(format("%s Initializing test %s %s%n%s", DASHES, testId, testCase, DASHES));
-
-                Object testInstance = InitCommand.class.getClassLoader().loadClass(testCase.getClassname()).newInstance();
-                bindProperties(testInstance, testCase, TestContainer.OPTIONAL_TEST_PROPERTIES);
-                TestContextImpl testContext = new TestContextImpl(testCase.id);
-                ProbesConfiguration probesConfiguration = parseProbeConfiguration(testCase);
-
-                tests.put(testContext.getTestId(),
-                        new TestContainer<TestContext>(testInstance, testContext, probesConfiguration, testCase));
-
-                if (serverInstance != null) {
-                    serverInstance.getUserContext().put(TestUtils.TEST_INSTANCE + ":" + testCase.id, testInstance);
-                }
-            } catch (Throwable e) {
-                log.fatal("Failed to init Test", e);
-                throw e;
-            }
-        }
-
-        public void process(StopCommand command) throws Exception {
-            try {
-                String testId = command.testId;
-                final String testName = "".equals(testId) ? "test" : testId;
-                TestContainer<TestContext> test = tests.get(command.testId);
-                if (test == null) {
-                    log.warn("Can't stop test, test with id " + command.testId + " does not exist");
-                    return;
-                }
-
-                log.info(format("%s %s.stop() %s", DASHES, testName, DASHES));
-                test.getTestContext().stop();
-            } catch (Exception e) {
-                log.fatal("Failed to execute test.stop", e);
-                throw e;
-            }
-        }
-
-        public boolean process(IsPhaseCompletedCommand command) throws Exception {
-            return !commands.containsKey(command.testId);
-        }
+    private void stop() {
+        log.info("Stopping threads...");
+        workerSocketProcessor.shutdown();
+        workerCommandRequestProcessor.shutdown();
     }
 
-    abstract class CommandThread extends Thread {
+    private HazelcastInstance createServerHazelcastInstance() throws Exception {
+        log.info("Creating Server HazelcastInstance");
 
-        private final Command command;
-        private final String testId;
+        XmlConfigBuilder configBuilder = new XmlConfigBuilder(hzFile);
+        Config config = configBuilder.build();
 
-        public CommandThread(Command command, String testId) {
-            this.command = command;
-            this.testId = testId;
-        }
-
-        public abstract void doRun() throws Throwable;
-
-        @Override
-        public final void run() {
-            try {
-                commands.put(testId, command);
-                doRun();
-            } catch (Throwable t) {
-                ExceptionReporter.report(testId, t);
-            } finally {
-                commands.remove(testId);
-            }
-        }
+        HazelcastInstance server = Hazelcast.newHazelcastInstance(config);
+        log.info("Successfully created Server HazelcastInstance");
+        return server;
     }
 
-    class TestContextImpl implements TestContext {
-        private final String testId;
-        volatile boolean stopped = false;
+    private HazelcastInstance createClientHazelcastInstance() throws Exception {
+        log.info("Creating Client HazelcastInstance");
 
-        TestContextImpl(String testId) {
-            this.testId = testId;
-        }
+        XmlClientConfigBuilder configBuilder = new XmlClientConfigBuilder(clientHzFile);
+        ClientConfig clientConfig = configBuilder.build();
 
-        @Override
-        public HazelcastInstance getTargetInstance() {
-            if (clientInstance != null) {
-                return clientInstance;
-            } else {
-                return serverInstance;
-            }
-        }
+        HazelcastInstance client = HazelcastClient.newHazelcastClient(clientConfig);
+        log.info("Successfully created Client HazelcastInstance");
+        return client;
+    }
 
-        @Override
-        public String getTestId() {
-            return testId;
+    private void signalStartToAgent(HazelcastInstance serverInstance) {
+        String address;
+        if (serverInstance != null) {
+            InetSocketAddress socketAddress = serverInstance.getCluster().getLocalMember().getInetSocketAddress();
+            address = socketAddress.getAddress().getHostAddress() + ":" + socketAddress.getPort();
+        } else {
+            address = "client:" + getHostAddress();
         }
-
-        @Override
-        public boolean isStopped() {
-            return stopped;
-        }
-
-        @Override
-        public void stop() {
-            stopped = true;
-        }
+        File file = new File("worker.address");
+        writeObject(address, file);
     }
 }
