+++ stabilizer/src/main/java/com/hazelcast/stabilizer/worker/TerminateWorkerException.java
@@ -1,5 +1,4 @@
 package com.hazelcast.stabilizer.worker;
 
 public class TerminateWorkerException extends RuntimeException {
-
 }
